pre-versions/v1/requirements.txt
streamlit>=1.35
pillow>=10.0
opencv-python-headless>=4.9
numpy>=1.26
pytest>=8.0
reportlab>=4.0